Rename plugin mocks to be more in line with hook names
