+++ test/mock/plugins/error.js
